feat (test): Add proper test harness.
There was no way to easily achieve integration tests. This new structure
should change this.
